Navigation Component Added.
