+++ app/build.gradle
@@ -38,6 +38,8 @@ android {
 
 dependencies {
 
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
     def lifecycle_version = "2.6.0-alpha03"
 
     // ViewModel
