+++ app/src/main/java/me/dwidar/movieapp/src/app/DetailsScreen/view/DetailsScreen.kt
@@ -1,0 +1,19 @@
+package me.dwidar.movieapp.src.app.DetailsScreen.view
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import me.dwidar.movieapp.R
+
+class DetailsScreen : Fragment()
+{
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_details_screen, container, false)
+    }
+}
