+++ app/src/main/res/layout/activity_main.xml
@@ -6,32 +6,23 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:layout_gravity="center"
-    tools:context=".src.app.MainScreen.view.MainActivity">
+    tools:context=".src.app.MainActivity">
 
     <include
         android:id="@+id/include"
         layout="@layout/main_action_bar" />
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/moviesGrid"
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/fragmentContainerView"
+        android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:layout_marginTop="@dimen/_16sdp"
-        android:numColumns="2"
-        app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+        app:defaultNavHost="true"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/include"
-        tools:listitem="@layout/movie_list_item" />
-
-    <ProgressBar
-        android:id="@+id/loadingBar"
-        android:layout_width="@dimen/_30sdp"
-        android:layout_height="@dimen/_30sdp"
-        app:layout_constraintBottom_toBottomOf="@+id/moviesGrid"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/moviesGrid" />
+        app:navGraph="@navigation/my_nav" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
